requirements.txt
protobuf==3.20.1
pillow==8.1.0
matplotlib==3.3.4
numpy==1.19.3
tqdm==4.56.0
requests==2.25.1
mediapipe==0.8.10.1
pandas==1.3.5
scikit-learn==1.0.2
pyrealsense2==2.50.0.3812
urllib3==1.26.9
wheel==0.37.1
opencv-contrib-python==4.5.5.62
PyYAML==6.0
PyQt5==5.15.7